abuse change typings func for dry
